Worst case O(V^2) to O(V+E) time complexity improvement for DFA minification // queue library change
In the breadth first search implementation in `_compute_reachable_states`, when processing the neighbours of a given vertex, for each vertex we check whether it is in `states_checked` and if not push it to the `states_to_check` queue, however, we only add each vertex to `states_checked` once we actually reach that vertex. This could lead to lots of duplication of vertices in the queue, with potentially a O(V * deg(V)) -> worst-case O(V^2) runtime.

Fixing this bug (in other words, adding to the visited set at the same time as we push the vertex to the queue) takes the runtime for minimising ~5000 fairly large DFAs (with around 6 states each after minimisation) from hours to a few minutes.

In addition, as it does not appear to be the case that we need thread-safe processing for the queues here, I changed the queue used from `queue.Queue` to `collections.deque` -- this also yields a performance improvement (see https://stackoverflow.com/a/20330499).