Use simpler regexes for NFA.shuffle_product() set laws test
Because of the exponential nature of NFA.shuffle_product(), and because
the use of longer regexes produces NFAs with more states, this single
test for set laws is remarkably slow (adding one whole second to the
test run time). To solve this, we can switch to using simpler regexes,
which eliminates the additional second while still testing the set laws
properly for this method.