Speed up Automata Creation
This commit speeds up the freezeValue method, which is the main
bottleneck of automata creation. There are two optimizations:

* Move the most frequent types up the if chain to hit it as early as
  possible
* Offer the option to disable the freezeValue call completely (the user
  must guarantee that the values are frozen or accept that the library
might break)

When benchmarking the creation of an automata accepting the language a*,
we see that disabling the freezeValue call yields a speedup of 3 times
if the values are not frozen and 2 times if the values are frozen.

The small overhead introduced by checking the ensure_frozen flag is
compensated by the reordering of if statements