Add blank_symbol to repr() for TMConfiguration
This is so that the TMConfiguration repr output is more closely aligned
to the __init__ signature for the class. We stil have the keyword
argument issue to solve, since the repr output currently lists all
parameters as positional arguments (whereas the __init__ signature
requires that blank_symbol and current_position be keyword-only).