First attempt at allow_partial
The semantics is defined such that if no transition is defined for a
certain symbol on a given state, we act as if there would be a
transition there leading to a trap state.

Reserve None to denote trap states.

The code is implemented such that there is little or no extra overhead
for complete DFAs. For partial DFAs, the code is implemented to perform
well with sparse DFAs. For most functions, the code will also perform
well with dense partial DFAs.