Fix repr() handling of sets within sets (fixes #150)
DFA.minify(retain_names=True) returns a DFA with a states set that is a
frozenset of frozensets. This was causing _get_repr_friendly_value() to
error out because the inner frozensets were being converted to sets,
which are not hashable, yet there were recursively being placed inside
of a set (which requires hashable members).

To solve this, we simply convert the outer frozenset to a set without
coercing the types of any of its elements.