Add missing dependencies to requirements.txt
The 'build' and 'tomli' packages should be explicitly present. This also
allows us to return to fully generating the requirements.txt file
completely automatically (previously, we could not do this because we
needed to install the [toml] addon with the coverage package, but we can
workaround this by simply installing the 'tomli' package directly).