`uv` improvements (#262)
* Add .venv to gitignore

.venv is the default virtualenv directory name for uv. It can be changed
at a user level by setting UV_PROJECT_ENVIRONMENT to the desired name,
however for those who don't set this variable, we still want to make
sure that Git ignores the default virtualenv.

* Use double quotes for strings in pyproject.toml

* Update Contributing guide for uv

* Do not lint against flake8-executable rules

I tend to add a shebang line to every file regardless of whether or not
it's actually executable.

* Format python.analysis.extraPaths

In VS Code settings.

* Remove -- from `uv run` examples

* Install enumerated Python version correctly on CI

* Remove obsolete code to configure pygraphviz

Now that we are using uv.